+++ cells/behaviors/broadcaster_test.go
@@ -27,7 +27,7 @@ import (
 // TestBroadcasterBehavior tests the broadcast behavior.
 func TestBroadcasterBehavior(t *testing.T) {
 	assert := audit.NewTestingAssertion(t, true)
-	env := cells.NewEnvironment("broadcast-behavior")
+	env := cells.NewEnvironment("broadcaster-behavior")
 	defer env.Stop()
 
 	env.StartCell("broadcast", behaviors.NewBroadcasterBehavior())